Specify bg and fg colors
The old code was rendering a white background with a user-defined foreground (i.e. text) color.

If the user has a dark theme, that means white background with white text, making the tool unusable.

This commit forces a pair of colors. It would be even beter to reuse the user-chosen colors from the current theme, but this is already good enough.